feat: separate server history env & Effects
