+++ packages/history/package.json
@@ -27,6 +27,7 @@
   "dependencies": {
     "@typed/common": "^2.1.0",
     "@typed/disposable": "^2.2.0",
+    "@typed/effects": "1.4.0",
     "@typed/env": "^3.3.0",
     "@typed/lambda": "^2.1.0",
     "@typed/new-type": "^2.1.0",
