+++ packages/history/source/pushState.ts
@@ -1,28 +1,20 @@
-import { Env } from '@typed/env'
-import { curry } from '@typed/lambda'
-import { Tuple } from '@typed/tuple'
+import { Effects, get } from '@typed/effects'
 import { HistoryEnv, Path } from './types'
 
 /**
  * Push state to history environment
  * @param data: A
  * @param path: Path
- * @returns Env<HistoryEnv<A>, Tuple<A, Location>>
+ * @returns Effects<HistoryEnv<A>, A>
  */
-export const pushState: {
-  <A>(data: A, path: Path): Env<HistoryEnv<A>, Tuple<A, Location>>
-  <A>(data: A): (path: Path) => Env<HistoryEnv<A>, Tuple<A, Location>>
-} = curry(
-  <A>(data: A, path: Path): Env<HistoryEnv<A>, Tuple<A, Location>> => ({
-    type: 'lazy',
-    runEnv: (f, { history, location }) => {
-      history.pushState(data, '', path)
+export function* pushState<A>(data: A, path: Path): Effects<HistoryEnv<A>, A> {
+  const { history } = yield* get<HistoryEnv<A>>()
 
-      return f([data, location])
-    },
-  }),
-)
+  history.pushState(data, '', path)
 
-export const pushPath: (path: Path) => Env<HistoryEnv<null>, Tuple<null, Location>> = pushState(
-  null,
-)
+  return data
+}
+
+export function* pushPath(path: Path): Effects<HistoryEnv, void> {
+  yield* pushState(null, path)
+}
