+++ packages/history/source/replaceState.ts
@@ -1,28 +1,12 @@
-import { Env } from '@typed/env'
-import { curry } from '@typed/lambda'
-import { Tuple } from '@typed/tuple'
+import { get } from '@typed/effects'
 import { HistoryEnv, Path } from './types'
 
-/**
- * Replace current state to history environment
- * @param data: A
- * @param path: Path
- * @returns Env<HistoryEnv<A>, Tuple<A, Location>>
- */
-export const replaceState: {
-  <A>(data: A, path: Path): Env<HistoryEnv<A>, Tuple<A, Location>>
-  <A>(data: A): (path: Path) => Env<HistoryEnv<A>, Tuple<A, Location>>
-} = curry(
-  <A>(data: A, path: Path): Env<HistoryEnv<A>, Tuple<A, Location>> => ({
-    type: 'lazy',
-    runEnv: (f, { history, location }) => {
-      history.replaceState(data, '', path)
+export function* replaceState<A>(data: A, path: Path) {
+  const { history } = yield* get<HistoryEnv<A>>()
 
-      return f([data, location])
-    },
-  }),
-)
+  history.replaceState(data, '', path)
 
-export const replacePath: (
-  path: Path,
-) => Env<HistoryEnv<null>, Tuple<null, Location>> = replaceState(null)
+  return data
+}
+
+export const replacePath = (path: Path) => replaceState<null>(null, path)
