+++ packages/history/source/server/createServerHistoryEnv.test.ts
@@ -1,18 +1,18 @@
 import { describe, given, it, Test } from '@typed/test'
 
 import { isBrowser } from '@typed/common'
-import { createHistoryEnv } from './createHistoryEnv'
+import { createServerHistoryEnv } from './createServerHistoryEnv'
 
 export const test: Test = describe(`createHistoryEnv`, [
   it(`returns Env with '/' as default location.pathname`, ({ equal }) => {
-    const { location } = createHistoryEnv()
+    const { location } = createServerHistoryEnv()
 
     equal('/', location.pathname)
   }),
 
   given(`an href`, [
     it(`returns { history: History, location: Location }`, ({ equal }) => {
-      const { location, history } = createHistoryEnv('http://www.example.com')
+      const { location, history } = createServerHistoryEnv('http://www.example.com')
 
       history.pushState(null, '', '/example')
 
@@ -34,7 +34,7 @@ export const test: Test = describe(`createHistoryEnv`, [
             return ok(true)
           }
 
-          const { location, history } = createHistoryEnv('https://www.example.com')
+          const { location, history } = createServerHistoryEnv('https://www.example.com')
 
           const href = 'http://www.example.com'
