+++ packages/history/source/server/index.ts
@@ -1,3 +1,4 @@
+export * from './createServerHistoryEnv'
 export * from './ServerHistory'
 export * from './ServerLocation'
 export * from './parseHref'
